Made the nano function not copy if file wasn't changed

+++ .config/fish/functions/nano.fish
@@ -1,5 +1,6 @@
 function nano --description 'Edit File with sudo if needed'
     # To anybody reading this function, I know it's fucking scuffed. Deal with it
+    set file ""
     set file $argv[1]
     set random_string (tr -dc A-Za-z0-9 </dev/urandom | head -c 13)
 
@@ -18,10 +19,14 @@ function nano --description 'Edit File with sudo if needed'
             mkdir -p $tmp_dir
             cp $file $tmp_dir
             kitten edit-in-kitty $tmp_dir/$basename
-            sudo cp $tmp_dir/$basename $file
-            sudo chmod $original_permissions $file
-            sudo chown $original_owner $file
-            sudo rm -rf $tmp_dir
+            if cmp -s $tmp_dir/$basename $file
+                echo "No changes made to $file"
+            else
+                sudo cp $tmp_dir/$basename $file
+                sudo chmod $original_permissions $file
+                sudo chown $original_owner $file
+            end
+            rm -rf $tmp_dir
         end
     end
 end
